feat: add task update method to TaskService and add JWT currentUsername validation.

+++ src/main/java/org/jakubmiczek/todoapp/service/TaskService.java
@@ -2,6 +2,8 @@ package org.jakubmiczek.todoapp.service;
 
 import org.jakubmiczek.todoapp.controller.dto.TaskRequest;
 import org.jakubmiczek.todoapp.controller.dto.TaskResponse;
+import org.jakubmiczek.todoapp.controller.dto.TaskUpdateRequest;
+import org.jakubmiczek.todoapp.exception.TaskAccessDeniedException;
 import org.jakubmiczek.todoapp.exception.TaskDoesNotExistException;
 import org.jakubmiczek.todoapp.exception.UserDoesNotExistException;
 import org.jakubmiczek.todoapp.entity.Task;
@@ -12,6 +14,7 @@ import org.jakubmiczek.todoapp.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TaskService {
@@ -38,9 +41,24 @@ public class TaskService {
         taskRepository.save(newTask);
     }
 
-    public void deleteTask(Long taskId) {
+    public void updateTask(TaskUpdateRequest taskUpdateRequest, String currentUsername) {
+        Task taskToUpdate = taskRepository.findById(taskUpdateRequest.taskId())
+                .orElseThrow(() -> new TaskDoesNotExistException(taskUpdateRequest.taskId()));
+
+        if(!taskToUpdate.getUser().getUsername().equals(currentUsername)) throw new TaskAccessDeniedException();
+
+        taskToUpdate.setTitle(taskUpdateRequest.title());
+        taskToUpdate.setDescription(taskUpdateRequest.description());
+        taskToUpdate.setStatus(taskUpdateRequest.status());
+        taskRepository.save(taskToUpdate);
+    }
+
+    public void deleteTask(Long taskId, String currentUsername) {
         Task task = taskRepository.findById(taskId)
                 .orElseThrow(() -> new TaskDoesNotExistException(taskId));
+
+        if(!task.getUser().getUsername().equals(currentUsername)) throw new TaskAccessDeniedException();
+
         taskRepository.delete(task);
     }
